refactor: Extract socket classes to separate module for size limits
Move SocketInterface and FakeSocket to core/worlds/soccer/socket_interface.py
to reduce rcssserver_adapter.py from 531 to 466 lines, meeting the 500-line
god class limit. Update imports in dependent files.

- Create new socket_interface.py with socket protocol and fake implementation
- Update rcssserver_adapter.py to import from socket_interface
- Update scripts/rcss_adapter_smoketest.py imports
- Update tests/fakes/fake_rcssserver.py imports

Fixes: tests/test_god_class_limits.py::test_no_new_god_classes
